+++ vbox/set_cloud_config.sh
@@ -1,6 +1,6 @@
 #!/usr/bin/env bash
 
-bosh -e vbox update-cloud-config ../bosh-deployment/virtualbox/cloud-config.yml \
+bosh -e vbox update-cloud-config ../bosh-deployment/warden/cloud-config.yml \
+-o ../operations/static-web.yml \
 -v internal_gw=192.168.50.1 \
--v internal_cidr=192.168.50.0/24 \
--v network_name=vboxnet0
+-v internal_cidr=192.168.50.0/24